move currentTransaction slice to transactions
