--- src/redux/slices/currentTransactionSlice/selector.ts
@@ -1,4 +1,0 @@
-import { RootState } from 'store/store'
-
-export const selectCurrentTransaction = ({ currentTransaction }: RootState) =>
-  currentTransaction
